requirements.txt
pandas 
mlflow==2.2.2
notebook
numpy
xgboost
scikit-learn
matplotlib
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
boto3
s3fs
-e .